Move CheckpointManager factories
