+++ ingestor/api/src/main/scala/za/co/absa/hyperdrive/ingestor/api/manager/OffsetManagerFactory.scala
@@ -15,11 +15,10 @@
  *   limitations under the License.
  */
 
-package za.co.absa.hyperdrive.ingestor.implementation.manager
+package za.co.absa.hyperdrive.ingestor.api.manager
 
 import org.apache.commons.configuration2.Configuration
-import za.co.absa.hyperdrive.ingestor.api.manager.OffsetManager
 
 trait OffsetManagerFactory {
-  def build(config: Configuration): OffsetManager
+  def apply(config: Configuration): OffsetManager
 }
