+++ ingestor/ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/manager/factories/OffsetManagerAbstractFactory.scala
@@ -19,8 +19,7 @@ package za.co.absa.hyperdrive.ingestor.implementation.manager.factories
 
 import org.apache.commons.configuration2.Configuration
 import org.apache.logging.log4j.LogManager
-import za.co.absa.hyperdrive.ingestor.api.manager.OffsetManager
-import za.co.absa.hyperdrive.ingestor.implementation.manager.OffsetManagerFactory
+import za.co.absa.hyperdrive.ingestor.api.manager.{OffsetManager, OffsetManagerFactory}
 import za.co.absa.hyperdrive.shared.utils.ClassLoaderUtils
 
 /**
@@ -37,6 +36,6 @@ object OffsetManagerAbstractFactory {
 
     val factoryName = config.getString(componentConfigKey)
     val factory = ClassLoaderUtils.loadSingletonClassOfType[OffsetManagerFactory](factoryName)
-    factory.build(config)
+    factory.apply(config)
   }
 }
